Move default exports to the top of component files
This ensures that the primary focus of the file can be identified
easily. A consequence of this is that components must be defined as a
`function` rather than a `const` so that they can be referenced
anywhere in the file due function hoisting.

This makes it easier to see at a glance whether a component is pure
and also makes the pure versions of the components available (for
tests).